requirements.txt
pytest==7.1.2
selenium==4.4.2
allure-pytest==2.10.0
pytest-xdist==2.5.0